refactor(vehicle): extract preset chip row out of service_reminder_section (Refs #563 phase: service_reminder_section)
Move the inline _PresetChipRow widget into a library-private
`part of` companion file. The section file shrinks from 309 to 273
LOC, back under the 300-line guidance for #563 without altering
behaviour: the public ServiceReminderSection API and the chip widget
visibility (still library-private via `part of`) are unchanged.
